refactor: Rename StyleControl to PropertyControl
- Since this control handles style and attributes, PropertyControl sound more generic
