+++ frontend/components.d.ts
@@ -102,6 +102,7 @@ declare module 'vue' {
     PlacementControl: typeof import('./src/components/PlacementControl.vue')['default']
     Play: typeof import('./src/components/Icons/Play.vue')['default']
     Plus: typeof import('./src/components/Icons/Plus.vue')['default']
+    PropertyControl: typeof import('./src/components/Controls/PropertyControl.vue')['default']
     PublishButton: typeof import('./src/components/PublishButton.vue')['default']
     RangeInput: typeof import('./src/components/Controls/RangeInput.vue')['default']
     Redirect: typeof import('./src/components/Icons/Redirect.vue')['default']
@@ -113,7 +114,6 @@ declare module 'vue' {
     Settings: typeof import('./src/components/Icons/Settings.vue')['default']
     SettingsGear: typeof import('./src/components/Icons/SettingsGear.vue')['default']
     StrikeThrough: typeof import('./src/components/Icons/StrikeThrough.vue')['default']
-    StyleControl: typeof import('./src/components/Controls/StyleControl.vue')['default']
     Switch: typeof import('./src/components/Controls/Switch.vue')['default']
     TabButtons: typeof import('./src/components/Controls/TabButtons.vue')['default']
     TemplatePagePreview: typeof import('./src/components/TemplatePagePreview.vue')['default']
