+++ frontend/src/components/BlockFlexLayoutHandler.vue
@@ -1,5 +1,5 @@
 <template>
-	<StyleControl
+	<PropertyControl
 		styleProperty="flexDirection"
 		defaultValue="row"
 		:component="OptionToggle"
@@ -8,14 +8,14 @@
 		:options="[
 			{ label: 'Horizontal', value: 'row', icon: 'arrow-right', hideLabel: true },
 			{ label: 'Vertical', value: 'column', icon: 'arrow-down', hideLabel: true },
-		]"></StyleControl>
-	<StyleControl
+		]"></PropertyControl>
+	<PropertyControl
 		styleProperty="alignItems"
 		label="Placement"
 		:enableStates="false"
 		v-if="blockController.isFlex()"
-		:component="PlacementControl"></StyleControl>
-	<StyleControl
+		:component="PlacementControl"></PropertyControl>
+	<PropertyControl
 		v-if="blockController.isFlex()"
 		styleProperty="justifyContent"
 		type="select"
@@ -25,13 +25,13 @@
 			{ label: 'Space Around', value: 'space-around' },
 			{ label: 'Space Evenly', value: 'space-evenly' },
 		]" />
-	<StyleControl
+	<PropertyControl
 		v-if="blockController.isFlex()"
 		label="Gap"
 		styleProperty="gap"
 		:enableSlider="true"
 		:unitOptions="['px', 'em', 'rem']" />
-	<StyleControl
+	<PropertyControl
 		styleProperty="flexWrap"
 		:component="OptionToggle"
 		label="Wrap"
@@ -40,9 +40,9 @@
 			{ label: 'No Wrap', value: 'nowrap' },
 			{ label: 'Wrap', value: 'wrap' },
 		]"
-		defaultValue="nowrap"></StyleControl>
+		defaultValue="nowrap"></PropertyControl>
 	<div class="flex flex-col gap-3" v-if="blockController.getParentBlock()?.isFlex()">
-		<StyleControl
+		<PropertyControl
 			label="Grow"
 			styleProperty="flexGrow"
 			:component="OptionToggle"
@@ -51,7 +51,7 @@
 				{ label: 'No', value: 0 },
 			]"
 			:defaultValue="0" />
-		<StyleControl
+		<PropertyControl
 			label="Shrink"
 			styleProperty="flexShrink"
 			:component="OptionToggle"
@@ -64,7 +64,7 @@
 </template>
 <script lang="ts" setup>
 import OptionToggle from "@/components/Controls/OptionToggle.vue";
-import StyleControl from "@/components/Controls/StyleControl.vue";
+import PropertyControl from "@/components/Controls/PropertyControl.vue";
 import blockController from "@/utils/blockController";
 import PlacementControl from "./PlacementControl.vue";
 </script>
